requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Werkzeug
flask_jwt_extended
